moved socket.emit events to backbone events
added a backbone events object:
  this.chatEvents = _.extend({},Backbone.Events);
this gets passed into this.chatView instance and initialized, allowing
vent.trigger inside of a view event to trigger
this.chatEvents.on('pushMessage'). pushMessage event hooks into sockChat
client and emits the socket.emit('chat message' ...) event.

the socket.on('') listening events are still listening outside of
backbone events by having the this.chatView instance passed into the
this.sockChat chat client. This is kind of ugly, and should be changed.